Switch navigation2D for custom AStarPath class
Note that this commit actually breaks targeting of mobs.
Will need to modify the get_distance method in actors/mob/mob.gd in order to fix it.

Also modified the actors/mob/mob.tscn to not have the children nodes be
centered. Needed for way the A* navigation was implemented.

Next step will be to use the tilemap_buildings instead of tilemap_ground
to do navigation, and reverse the .has_cell check for determining the
path, making sure that the path cannot extend further than the region of
the tilemap.

An easy way to do that is add a "map_size" export variable in a
script attached to the tilemap_buildings, stating the maximum size the
tilemap is allow to extend (or rather, how far away from the
tilemap_buildings node the A* algorithm will check to add to the AStarPath's path array).