Further separate tilemaps by usage
- background: Stuff in the background. Mostly grass, but could be
  fancier grasses sometimes.

- foreground: Anything on top of the background, mostly dirt path and
  fences for now.

- zone_path: Tilemap for A* pathfinding. Any used_cells will be added to
  the AStarPath walkable_cells if they do not have an obstacle from
  player_buildings on the same cell.

- zone_build: Tilemap for buildeable area. Any used_cells will be
  considered valid for the player to maze upon. Basically the zone_path
  without the padding on the sides, but done manually to allow
  flexibility.

- player_buildings: Tilemap for the actual buildings placed by the
  player during build mode. Used in the A* Pathfinding as the
  obstacles_array for calculating path. If this blocks the way to the
  path, the mobs will not advance (and eventually will have a way to
  automatically detect it and prevent placing the building.

Also took some time to fill the map with tiles and place the areas for
what will eventually be the storage (4x3 zone, top left) and the
nestboxes area (3x3, to the right of the storage).

Set start_position to the mob_spawn (and placed it further back right),
and the end_position to the egg storage (not sure what will happen when
they'll get there but it seemed to be the destination we wanted to have
initially so we'll go for that for now.)

Some other stuff I'm sure of it.